Backend/requirements.txt
# requirements.txt

annotated-types==0.6.0
anyio==4.3.0
exceptiongroup==1.2.0
fastapi==0.109.2
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
